+++ paddle/fluid/CMakeLists.txt
@@ -6,5 +6,9 @@ add_subdirectory(imperative)
 add_subdirectory(operators)
 add_subdirectory(pybind)
 add_subdirectory(eager)
+cmake_policy(SET CMP0079 NEW)
+target_link_libraries(test_tracer dygraph_function)
+target_link_libraries(test_hooks dygraph_function)
+cmake_policy(SET CMP0079 OLD)
 # NOTE: please add subdirectory inference at last.
 add_subdirectory(inference)
